requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.3
Flask-Cors==4.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
numpy==1.26.4
paramiko==3.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
scipy==1.13.0
urllib3==2.2.1
Werkzeug==3.0.2
